requirements.txt
# Dev environment
pip
autopep8

# App
flask
firebase
pyrebase4
firebase-admin
python-dotenv
